Fix a freeze in the oven code.
Previously, a freeze would occur when cooking finished on an item which
already had a full stack in the oven output.  The slot # to use would
initially be set to the full one and then would get into the while loop
which attempts to split up the outputting items and get into an infinite
loop because its chosen stack was full and the while loop can't handle
that.

This prevents the code from committing to a full output slot.