Prevent nulls in recipes from derailing the mod.
Nulls are frequently used to provide spacing in shaped recipes.  Currently
foods with nulls in the recipe crash out when they're shown in the
crafting book/table.  The problem is that because the food ingredient list
is an array that's pre-initialized there's not really a good way of
removing the nulls without pre-searching for the ingredient count.  It's
just easier to use the arraylist instead of an actual array, and it
doesn't really cost more except during the cache building process which is
already lengthy.  So I converted the ingredient cache to arraylists and
prevented nulls from being inserted into ingredient lists.