Fix tools better.
The flint cutter is a bit weird because it's finite, so we need the
new method.  The other recipes get picked up as food, but the reason
why they don't show up in the book has nothing to do with wildcards
and everything to do with the book not understanding stuff with
metadata, which I'm gonna fix.